requirements.txt
pycryptodome
pycryptodomex
tabulate
getpass
